requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-debug-toolbar==4.3.0
django-filter==24.2
djangorestframework==3.14.0
drf-spectacular==0.26.5
drf_spectacular_extensions==0.0.2
idna==3.4
inflection==0.5.1
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
Pillow==10.1.0
psycopg==3.1.18
psycopg2-binary==2.9.9
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
rpds-py==0.12.0
six==1.16.0
sqlparse==0.4.4
telebot==0.0.5
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.0.7
schedule==1.2.1
